requirements.txt
trino==0.328.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
python-dotenv==1.0.0